Refactor: Remove unused enhanced integration modules
These modules are no longer needed as their functionality has been integrated elsewhere.
